backend/requirements.txt
dotenv
google-genai
google-api-python-client